[fix] Ensure that feed ID are 32-bit integers (close #25) -
There was a bug where in some places we referd to `int` instead of
`int32_t`, which is the same on most 32-bit architectures. This
came up trying to compile for the MSP430 (a 16-bit architecture).
So the actual Xively feed IDs are unsigned 32-bit integers, let's
make sure we have that right and define a specific type for it too.